+++ src/Instagram.php
@@ -26,6 +26,13 @@ use Http\Message\RequestFactory;
 class Instagram
 {
     /**
+     * The access token.
+     *
+     * @var string
+     */
+    protected $accessToken;
+
+    /**
      * The http client.
      *
      * @var \Http\Client\HttpClient
@@ -42,13 +49,15 @@ class Instagram
     /**
      * Create a new instagram instance.
      *
+     * @param string $accessToken
      * @param \Http\Client\HttpClient|null $httpClient
      * @param \Http\Message\RequestFactory|null $requestFactory
      *
      * @return void
      */
-    public function __construct(HttpClient $httpClient = null, RequestFactory $requestFactory = null)
+    public function __construct(string $accessToken, HttpClient $httpClient = null, RequestFactory $requestFactory = null)
     {
+        $this->accessToken = $accessToken;
         $this->httpClient = $httpClient ?: HttpClientDiscovery::find();
         $this->requestFactory = $requestFactory ?: MessageFactoryDiscovery::find();
     }
@@ -56,24 +65,24 @@ class Instagram
     /**
      * Fetch the media items.
      *
-     * @param string $user
-     *
      * @throws \Vinkla\Instagram\InstagramException
      *
      * @return array
      */
-    public function get(string $user): array
+    public function get(): array
     {
-        $uri = sprintf('https://www.instagram.com/%s/media/', $user);
+        $uri = sprintf('https://api.instagram.com/v1/users/self/media/recent?access_token=%s', $this->accessToken);
 
         $request = $this->requestFactory->createRequest('GET', $uri);
 
         $response = $this->httpClient->sendRequest($request);
 
-        if ($response->getStatusCode() === 404) {
-            throw new InstagramException(sprintf('The user [%s] wasn\'t found.', $user));
+        if ($response->getStatusCode() === 400) {
+            $body = json_decode((string) $response->getBody());
+
+            throw new InstagramException($body->meta->error_message);
         }
 
-        return json_decode((string) $response->getBody())->items;
+        return json_decode((string) $response->getBody())->data;
     }
 }