+++ tests/InstagramTest.php
@@ -13,6 +13,8 @@ declare(strict_types=1);
 
 namespace Vinkla\Tests\Instagram;
 
+use GuzzleHttp\Psr7\Response;
+use Http\Mock\Client;
 use PHPUnit\Framework\TestCase;
 use Vinkla\Instagram\Instagram;
 use Vinkla\Instagram\InstagramException;
@@ -26,16 +28,25 @@ class InstagramTest extends TestCase
 {
     public function testGet()
     {
-        $items = (new Instagram())->get('jerryseinfeld');
+        $response = new Response(200, [], json_encode([
+            'data' => range(0, 19),
+        ]));
+
+        $client = new Client();
+        $client->addResponse($response);
+
+        $instagram = new Instagram('jerryseinfeld', $client);
+        $items = $instagram->get();
 
         $this->assertInternalType('array', $items);
         $this->assertCount(20, $items);
     }
 
-    public function testNotFound()
+    public function testError()
     {
         $this->expectException(InstagramException::class);
+        $this->expectExceptionMessage('The access_token provided is invalid.');
 
-        (new Instagram())->get('imspeechlessihavenospeech');
+        (new Instagram('imspeechlessihavenospeech'))->get();
     }
 }